Use Alice swap factory in production
