+++ swap/tests/happy_path_restart_bob_before_comm.rs
@@ -7,7 +7,7 @@ pub mod testutils;
 
 #[tokio::test]
 async fn given_bob_restarts_after_xmr_is_locked_resume_swap() {
-    testutils::init(|test| async move {
+    testutils::setup_test(|test| async move {
         let alice_swap = test.new_swap_as_alice().await;
         let bob_swap = test.new_swap_as_bob().await;
