+++ swap/tests/punish.rs
@@ -9,7 +9,7 @@ pub mod testutils;
 /// the encsig and fail to refund or redeem. Alice punishes.
 #[tokio::test]
 async fn alice_punishes_if_bob_never_acts_after_fund() {
-    testutils::init(|test| async move {
+    testutils::setup_test(|test| async move {
         let alice_swap = test.new_swap_as_alice().await;
         let bob_swap = test.new_swap_as_bob().await;
